Added CI-CD part to configure.sh file for automation

+++ modules/sc-mesh-secure-deployment/configure.sh
@@ -55,6 +55,7 @@ function get_SSID() {
 }
 
 #-----------------------------------------------------------------------------#
+#TODO: Predefined choice for all cases?
 function ap_connect {
 echo '> Please choose from the list of available interfaces...'
 interfaces_arr=($(ip link | awk -F: '$0 !~ "lo|vir|doc|eth|bat|^[^0-9]"{print $2}'))
@@ -65,13 +66,25 @@ get_SSID <<< "$(iw $sta_if scan)"
 echo '> scanning available Access Point, select ssid from scan list...'
 menu_from_array "${ssid_array[@]}"
 ssid=$choice
-read -p "- Password: " password
-cat <<EOF > conf/ap.conf
-network={
-  ssid="$ssid"
-  psk="$password"
-}
+#ssid should be $ssid in conf/ap.conf
+if [[ "$2" == "-ci" ]]; then
+  password = "root"
+  echo "- Password: $password"
+  cat <<EOF > conf/ap.conf
+  network={
+    ssid="WirelessLab" 
+    psk="$password"
+  }
 EOF
+else
+  read -p "- Password: " password
+  cat <<EOF > conf/ap.conf
+  network={
+    ssid="WirelessLab"
+    psk="$password"
+  }
+EOF
+fi
 
 #mesh(IBSS) and sta interface are not supported concurrently.
 #Stop mesh(IBSS) service on the selected vif, if already running.
@@ -311,9 +324,14 @@ function client {
     popd
   fi
   # Connect to the same AP as the server
-  read -p "> We need to be connect to the same network as the server... Connect to an Access Point? (Y/N): " confirm
-  if [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]]; then
-    ap_connect
+  # Add if statement for CI/CD
+  if [[ "$2" == "-ci" ]]; then
+    echo "We need to be connect to the same network as the server... Connect to an Access Point? (Y/N): No"
+  else
+    read -p "> We need to be connect to the same network as the server... Connect to an Access Point? (Y/N): " confirm
+    if [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]]; then
+      ap_connect $1 $2
+    fi
   fi
   echo -n '> Server discovery...'
   # get server IPv4 and hostname
@@ -325,14 +343,22 @@ function client {
   server_ip=${server_details[0]}
   server_host=${server_details[1]}
   echo "> We will use src/ecc_key.der if it already exists, or we can try and fetch it..."
-  read -p "> Do you want to fetch the certificate from the server $server_host@$server_ip? (Y/N): " confirm
-  if [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]]; then
+if [[ "$2" == "-ci"  ]]; then
+    server_user = "root"
+    echo "> Do you want to fetch the certificate from the server $server_host@$server_ip? (Y/N): Yes"
     echo '> Fetching certificate from server...'
-    read -p "- Server Username: " server_user
+    echo "- Server Username: " $server_user
     # pull the key from the server
     scp $server_user@$server_ip:/opt/container-data/mesh/mesh_com/modules/sc-mesh-secure-deployment/src/ecc_key.der $MESH_COM_ROOT$KEY_PATH
+  else
+    read -p "> Do you want to fetch the certificate from the server $server_host@$server_ip? (Y/N): " confirm
+    if [[ $confirm == [yY] || $confirm == [yY][eE][sS] ]]; then
+      echo '> Fetching certificate from server...'
+      read -p "- Server Username: " server_user
+      # pull the key from the server
+      scp $server_user@$server_ip:/opt/container-data/mesh/mesh_com/modules/sc-mesh-secure-deployment/src/ecc_key.der $MESH_COM_ROOT$KEY_PATH
+    fi
   fi
-
   echo '> Configuring the client and connecting to server...'
   if [ "$2" == "-t" ] ; then
     python3 $MESH_COM_ROOT$CLIENT_SRC_PATH -c $MESH_COM_ROOT$KEY_PATH -s http://$server_ip:5000 --test -m $mesh_mode
